+++ src/ConnectionConfig.test.tsx
@@ -29,7 +29,7 @@ const makeJsonData: (
     clientEmail: 'test@grafana.com',
     tokenUri: 'https://accounts.google.com/o/oauth2/token',
     defaultProject: 'test-project',
-    oauthPassThru: authenticationType === GoogleAuthType.OAUTH_PASSTHROUGH,
+    oauthPassThru: authenticationType === GoogleAuthType.ForwardOAuthIdentity,
   });
 };