+++ src/components/index.ts
@@ -1,4 +1,5 @@
 export * from './AuthConfig';
 export * from './JWTConfigEditor';
 export * from './JWTForm';
+export * from './OAuthPassthroughConfigEditor';
 export * from './WIFConfigEditor';