+++ src/types.ts
@@ -4,6 +4,7 @@ export const GoogleAuthType = {
   JWT: 'jwt',
   GCE: 'gce',
   WIF: 'workloadIdentityFederation',
+  ForwardOAuthIdentity: 'forwardOAuthIdentity',
 } as const;
 
 export interface DataSourceOptions extends DataSourceJsonData {
@@ -16,6 +17,7 @@ export interface DataSourceOptions extends DataSourceJsonData {
   usingImpersonation?: boolean;
   workloadIdentityPoolProvider?: string;
   wifServiceAccountEmail?: string;
+  oauthPassThru?: boolean;
 }
 
 export interface DataSourceSecureJsonData {